tests/test_requirements.txt
requests==2.18.0
urllib3==1.21.1
PyJWT==1.5.3
marshmallow==3.0.0b7